% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{absolute_copy_number_distance}
\alias{absolute_copy_number_distance}
\title{Compute distance function for fitting relative copy numbers to absolute}
\usage{
absolute_copy_number_distance(
  ploidy,
  cellularity,
  relative_copy_numbers,
  weights = NULL,
  distance_function = c("MAD", "RMSD")
)
}
\arguments{
\item{ploidy}{the tumour ploidy.}

\item{cellularity}{the cellularity, i.e. the fraction of cells that are from
the tumour.}

\item{relative_copy_numbers}{a numeric vector containing relative copy
numbers, i.e. ratios of copy numbers to the average copy number.}

\item{weights}{a numeric vector of weights to apply to each copy number value
(should be same length as relative_copy_numbers)}

\item{distance_function}{the distance function to use, either "MAD" for the
mean absolute difference or "RMSD" for the root mean square difference, where
differences are between the fitted absolute copy number values and the
nearest whole number.}
}
\value{
the distance in the fitted absolute copy numbers to whole numbers.
}
\description{
Computes a distance function on fitting relative copy numbers to whole number
absolute values with the given ploidy and cellularity.
}
\examples{
data(copy_number)
copy_number <- copy_number[copy_number$sample == "X17222", ]
absolute_copy_number_distance(3, 0.67, copy_number$segmented)
}
