% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{absolute_copy_number_distance_grid}
\alias{absolute_copy_number_distance_grid}
\title{Compute distance function for fitting relative copy numbers for a grid of
ploidies and cellularities}
\usage{
absolute_copy_number_distance_grid(
  relative_copy_numbers,
  weights = NULL,
  min_ploidy = 1.5,
  max_ploidy = 5.5,
  ploidy_step = 0.01,
  min_cellularity = 0.01,
  max_cellularity = 1,
  cellularity_step = 0.01,
  distance_function = c("MAD", "RMSD")
)
}
\arguments{
\item{relative_copy_numbers}{a numeric vector containing relative copy
numbers, i.e. ratios of copy numbers to the average copy number.}

\item{weights}{a numeric vector of weights to apply to each copy number value
(should be same length as relative_copy_numbers)}

\item{min_ploidy, max_ploidy}{the range of ploidies.}

\item{ploidy_step}{the stepwise increment of ploidies along the grid.}

\item{min_cellularity, max_cellularity}{the range of cellularities.}

\item{cellularity_step}{the stepwise increment of cellularities along the
grid.}

\item{distance_function}{the distance function to use, either "MAD" for the
mean absolute difference or "RMSD" for the root mean square difference, where
differences are between the fitted absolute copy number values and the
nearest whole number.}
}
\value{
the distance in the fitted absolute copy numbers to whole numbers.
}
\description{
Computes a distance function on fitting relative copy numbers to whole number
absolute values for a grid of ploidies and cellularities.
}
\examples{
data(copy_number)
copy_number <- copy_number[copy_number$sample == "X17222", ]
segments <- copy_number_segments(copy_number)
distances <- absolute_copy_number_distance_grid(segments$copy_number, segments$weight)
}
