% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{absolute_copy_number_distance_heatmap}
\alias{absolute_copy_number_distance_heatmap}
\title{Distance heat map for fitting a set of relative copy numbers}
\usage{
absolute_copy_number_distance_heatmap(
  relative_copy_numbers,
  weights = NULL,
  distance_function = "MAD",
  min_ploidy = 1.5,
  max_ploidy = 5.5,
  low_distance_colour = "red",
  high_distance_colour = "blue",
  xlabel = "cellularity",
  ylabel = "ploidy"
)
}
\arguments{
\item{relative_copy_numbers}{a numeric vector containing relative copy
numbers, i.e. ratios of copy numbers to the average copy number.}

\item{weights}{a numeric vector of weights to apply to each copy number value
(should be same length as relative_copy_numbers)}

\item{distance_function}{the distance function to use, either "MAD" for the
mean absolute difference or "RMSD" for the root mean square difference, where
differences are between the fitted absolute copy number values and the
nearest whole number.}

\item{min_ploidy}{the minimum ploidy to display.}

\item{max_ploidy}{the minimum ploidy to display.}

\item{low_distance_colour}{the colour for low distances.}

\item{high_distance_colour}{the colour for high distances.}

\item{xlabel, ylabel}{x- and y-axis labels.}
}
\value{
a \code{ggplot} object
}
\description{
Heat map of the distance of the given relative copy numbers to whole numbers
after scaling to absolute copy numbers for various ploidies and
cellularities.
}
\examples{
data(copy_number)
copy_number <- copy_number[copy_number$sample == "X17222", ]
segments <- copy_number_segments(copy_number)

absolute_copy_number_distance_heatmap(
  segments$copy_number,
  segments$weight,
  distance_function = "MAD"
)
}
