% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{absolute_to_relative_copy_number}
\alias{absolute_to_relative_copy_number}
\title{Convert absolute copy numbers to relative copy numbers}
\usage{
absolute_to_relative_copy_number(absolute_copy_numbers, ploidy, cellularity)
}
\arguments{
\item{absolute_copy_numbers}{a numeric vector of absolute copy numbers.}

\item{ploidy}{the tumour ploidy.}

\item{cellularity}{the cellularity, i.e. the fraction of cells that are from
the tumour.}
}
\value{
a numeric vector containing relative copy numbers, i.e. ratios of
copy numbers to the average copy number.
}
\description{
Convert absolute copy numbers to relative copy numbers based on the given
ploidy and cellularity.
}
\examples{
absolute_to_relative_copy_number(4, 4.01, 0.77)
absolute_to_relative_copy_number(1:10, 4.01, 0.77)
}
