% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{chromosome_copy_number_plot}
\alias{chromosome_copy_number_plot}
\title{Copy number plot for a single chromosome}
\usage{
chromosome_copy_number_plot(
  copy_number,
  segments,
  chromosome,
  start = NULL,
  end = NULL,
  copy_number_steps = NULL,
  genes = NULL,
  max_points_to_display = Inf,
  min_copy_number = NULL,
  max_copy_number = NULL,
  copy_number_breaks = NULL,
  point_colour = "black",
  point_alpha = NULL,
  point_size = NULL,
  segment_colour = "red",
  segment_alpha = 1,
  segment_line_size = 0.75,
  copy_number_step_colour = "blue",
  copy_number_step_alpha = 0.35,
  copy_number_step_line_size = 0.75,
  gene_colour = "darkgreen",
  gene_alpha = 0.25,
  position_scale = 1e-06,
  xlabel = "position (Mbp)",
  ylabel = "copy number"
)
}
\arguments{
\item{copy_number}{a data frame containing \code{chromosome}, \code{position}
and \code{copy_number} columns with a row for each copy number bin.}

\item{segments}{a data frame containing \code{chromosome},
\code{start}, \code{end} and \code{copy_number} columns.}

\item{chromosome}{the chromosome to display.}

\item{start}{the start coordinate within the specified chromosome.}

\item{end}{the end coordinate within the specified chromosome.}

\item{copy_number_steps}{a data frame containing \code{absolute_copy_number}
and \code{copy_number} columns.}

\item{genes}{a data frame containing \code{name}, \code{chromosome},
\code{start} and \code{end} columns.}

\item{max_points_to_display}{maximum number of copy number points to display
(downsampling may be carried out if there are more copy number values than
this number).}

\item{min_copy_number}{the minimum \code{copy_number} to display.}

\item{max_copy_number}{the maximum \code{copy_number} to display.}

\item{copy_number_breaks}{breaks at which grid lines will be displayed.}

\item{point_colour}{the colour of the copy number points.}

\item{point_alpha}{the transparency of the copy number points.}

\item{point_size}{= the size of the copy number points.}

\item{segment_colour}{the colour of the copy number segments.}

\item{segment_alpha}{the transparency of the copy number segments.}

\item{segment_line_size}{the size of the lines for copy number segments.}

\item{copy_number_step_colour}{the colour of the copy number step lines.}

\item{copy_number_step_alpha}{the transparency of the copy number step lines.}

\item{copy_number_step_line_size}{the size of the lines for the copy number
steps.}

\item{gene_colour}{the colour of the gene bars.}

\item{gene_alpha}{the transparency of the gene bars.}

\item{position_scale}{the scaling factor for the position x-axis.}

\item{xlabel, ylabel}{x- and y-axis labels.}
}
\value{
a \code{ggplot} object
}
\description{
Chromosome copy number plot with gene annotation.
}
\details{
\code{copy_number} values can be on the relative or absolute scales or can
be log2 ratios but the same scale should be used consistently in each of the
data frames.
}
\examples{
data(copy_number)
data(genes)

copy_number <- copy_number[copy_number$sample == "X17222", ]
copy_number$position <- round((copy_number$start + copy_number$end) / 2)
segments <- copy_number_segments(copy_number)

absolute_copy_numbers <- 0:8
relative_copy_numbers <- absolute_to_relative_copy_number(absolute_copy_numbers, ploidy = 4.01, cellularity = 0.81)
copy_number_steps <- data.frame(absolute_copy_number = absolute_copy_numbers, copy_number = relative_copy_numbers)

chromosome_copy_number_plot(
  copy_number,
  segments,
  chromosome = 3,
  copy_number_steps = copy_number_steps,
  genes = genes,
  min_copy_number = 0.25, max_copy_number = 2.5)

# display copy number as log2 ratios
copy_number$copy_number <- log2(copy_number$copy_number)
segments$copy_number <- log2(segments$copy_number)
copy_number_steps$copy_number <- log2(copy_number_steps$copy_number)

chromosome_copy_number_plot(
  copy_number,
  segments,
  chromosome = 17, start = 7250000, end = 7750000,
  copy_number_steps = copy_number_steps,
  genes = genes,
  min_copy_number = -2, max_copy_number = 2,
  position_scale = 1,
  xlabel = "position"
)

}
