% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chromosome_lengths}
\alias{chromosome_lengths}
\title{Obtain chromosome lengths}
\usage{
chromosome_lengths(copy_number)
}
\arguments{
\item{copy_number}{a data frame containing \code{chromosome} and \code{end}
columns.}
}
\value{
a data frame containing \code{chromosome} and \code{length} columns.
}
\description{
Obtains chromosome lengths based on the bin coordinates within the given copy
number data. Used in the genome copy number plot.
}
\examples{
data(copy_number)
chromosome_lengths(copy_number)
}
