% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chromosome_offsets}
\alias{chromosome_offsets}
\title{Obtain chromosome lengths and offsets}
\usage{
chromosome_offsets(copy_number)
}
\arguments{
\item{copy_number}{a data frame containing \code{chromosome} and \code{end}
columns.}
}
\value{
a data frame containing \code{chromosome}, \code{length} and
\code{offset} columns.
}
\description{
Obtains chromosome lengths and offset for the given copy number data.
}
\details{
Offsets are used to convert chromosome coordinates into genomic coordinates.
}
\examples{
data(copy_number)
chromosome_offsets(copy_number)
}
