% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_genomic_coordinates}
\alias{convert_to_genomic_coordinates}
\title{Convert chromosome coordinates to genomic coordinates}
\usage{
convert_to_genomic_coordinates(copy_number, column_names, offsets = NULL)
}
\arguments{
\item{copy_number}{a data frame containing \code{chromosome} and one of more
coordinate columns, e.g. \code{position}, \code{start} or \code{end}.}

\item{column_names}{a vector of column names to convert to genomic
coordinates.}

\item{offsets}{a data frame containing \code{chromosome} and \code{offset}
columns (optional, if not given then offsets will be calculated for the
given copy number).}
}
\description{
Converts the specified columns from chromosome coordinates to genomic
coordinates.
}
\details{
Genomic coordinates are obtained by adding the chromosome coordinate to the
sum of the lengths of all preceding chromosomes.
}
\examples{
data(copy_number)
copy_number <- copy_number[copy_number$sample == "X17222", ]
copy_number <- convert_to_genomic_coordinates(copy_number, c("start", "end"))
}
