% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{copy_number}
\alias{copy_number}
\title{Sample copy number data set}
\format{
A data frame with 206398 rows and 6 variables:
\describe{
  \item{sample}{the name of the sample}
  \item{chromosome}{the chromosome}
  \item{start}{the start position of the bin within the chromosome}
  \item{end}{the end position of the bin within the chromosome}
  \item{copy_number}{the relative copy number for the bin}
  \item{segmented}{the segmented relative copy number for the bin}
}
}
\usage{
copy_number
}
\description{
A dataset containing copy number data for 2 samples obtained through shallow
whole genome sequencing and processing using 30kb bins.
}
\details{
Each row represents one bin, i.e. a window of 30 kilobases. The dataset
contains both pre- and post-segmented copy number values. These are relative
copy numbers, i.e. ratios of copy numbers to the average copy number, and
have not been log2-transformed.
}
\keyword{datasets}
