% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_number_density}
\alias{copy_number_density}
\title{Copy number density estimation}
\usage{
copy_number_density(
  copy_numbers,
  min_copy_number = NULL,
  max_copy_number = NULL,
  n = 512
)
}
\arguments{
\item{copy_numbers}{a numeric vector of relative copy number values.}

\item{min_copy_number}{the lower copy number value in the range of values to
estimate the density.}

\item{max_copy_number}{the upper copy number value in the range of values to
estimate the density.}

\item{n}{the number of equally-spaced points for which the density will be
estimated (a smaller number may be returned if \code{min_copy_number} and/or
\code{max_copy_number} are specified).}
}
\value{
a data frame with copy number and density columns
}
\description{
Obtain density estimates for the distribution of the given copy number
values.
}
\examples{
data(copy_number)
copy_number <- copy_number[copy_number$sample == "X17222", ]

density <- copy_number_density(copy_number$segmented)

density <- copy_number_density(copy_number$segmented,
                               min_copy_number = 0,
                               max_copy_number = 2.5)
}
