% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{copy_number_density_plot}
\alias{copy_number_density_plot}
\title{Copy number density plot}
\usage{
copy_number_density_plot(
  copy_numbers,
  copy_number_steps = NULL,
  min_copy_number = NULL,
  max_copy_number = NULL,
  copy_number_step_colour = "blue",
  copy_number_step_alpha = 0.35,
  copy_number_step_line_size = 0.75,
  xlabel = "copy number",
  ylabel = "density"
)
}
\arguments{
\item{copy_numbers}{a numeric vector containing copy number values.}

\item{copy_number_steps}{a data frame containing \code{copy_number}
and \code{absolute_copy_number} columns.}

\item{min_copy_number}{the minimum relative copy number to display.}

\item{max_copy_number}{the maximum relative copy number to display.}

\item{copy_number_step_colour}{the colour of the copy number step lines.}

\item{copy_number_step_alpha}{the transparency of the copy number step lines.}

\item{copy_number_step_line_size}{the size of the lines for the copy number
steps.}

\item{xlabel, ylabel}{x- and y-axis labels.}
}
\value{
a \code{ggplot} object
}
\description{
Creates a ggplot for the density of the given copy numbers values.
}
\examples{
data(copy_number)
copy_number <- copy_number[copy_number$sample == "X17222", ]

copy_number_density_plot(copy_number$segmented)

absolute_copy_numbers <- 0:8
relative_copy_numbers <- absolute_to_relative_copy_number(absolute_copy_numbers, ploidy = 4.01, cellularity = 0.81)
copy_number_steps <- data.frame(absolute_copy_number = absolute_copy_numbers, copy_number = relative_copy_numbers)

copy_number_density_plot(copy_number$segmented, copy_number_steps, 0, 2.5)
}
