% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_number_maxima}
\alias{copy_number_maxima}
\title{Obtain maxima in a distribution of relative copy numbers.}
\usage{
copy_number_maxima(
  copy_numbers,
  min_copy_number = 0,
  max_copy_number = 3,
  lower_threshold = 0
)
}
\arguments{
\item{copy_numbers}{a numeric vector of absolute copy number values.}

\item{min_copy_number}{the lower copy number value in the range of
values to estimate the density.}

\item{max_copy_number}{the upper copy number value in the range of
values to estimate the density.}

\item{lower_threshold}{the lower threshold for a maximum as a fraction of
the density of the largest maximum.}
}
\value{
a data frame with copy number and density columns and a row for each
maximum in the distribution.
}
\description{
This function obtains peaks in the density for the given set of relative
copy numbers.
}
\examples{
data(copy_number)
copy_number <- copy_number[copy_number$sample == "X17222", ]
maxima <- copy_number_maxima(copy_number$segmented, lower_threshold = 0.1)
}
