% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{copy_number_segments}
\alias{copy_number_segments}
\title{Collapse copy number bins into segments}
\usage{
copy_number_segments(copy_number)
}
\arguments{
\item{copy_number}{a data frame containing \code{sample}, \code{chromosome},
\code{start}, \code{end} and \code{segmented} columns.}
}
\value{
a data frame containing copy number segments with
\code{segment}, \code{sample}, \code{chromosome}, \code{start}, \code{end},
\code{copy_number}, \code{bin_count}, \code{sum_of_bin_lengths} and
\code{weight}.
}
\description{
Collapses copy number segments for the given copy number data frame.
}
\details{
Adjacent bins with the same segmented copy number (\code{segmented} column)
for the same sample and on the same chromosome are treated as belonging to
the same segment.

The returned data frame contains a row for each segment with columns for
the segment number (\code{segment}), the number of bins or rows in the
provided copy number data frame that were collapsed to form the segment
(\code{bin_count}), the sum of lengths of those bins
(\code{sum_of_bin_lengths}) and the weight for the segment \code{weight}).
The segment weight is the median-scaled sum of bin lengths.
}
\examples{
data(copy_number)

# segments for all samples
segments <- copy_number_segments(copy_number)

# segments for single sample
copy_number <- copy_number[copy_number$sample == "X17222", ]
segments <- copy_number_segments(copy_number)
}
