% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{distance_heatmap}
\alias{distance_heatmap}
\title{Heat map for given distances for grid of ploidies and cellularities}
\usage{
distance_heatmap(
  distances,
  low_distance_colour = "red",
  high_distance_colour = "blue",
  xlabel = "cellularity",
  ylabel = "ploidy"
)
}
\arguments{
\item{distances}{a data frame containing \code{ploidy}, \code{cellularity}
and \code{distance} columns.}

\item{low_distance_colour}{the colour for low distances.}

\item{high_distance_colour}{the colour for high distances.}

\item{xlabel, ylabel}{x- and y-axis labels.}
}
\value{
a \code{ggplot} object
}
\description{
Heat map of the given distances for a grid of ploidy and cellularity
values.
}
\examples{
data(copy_number)
copy_number <- copy_number[copy_number$sample == "X17222", ]
segments <- copy_number_segments(copy_number)

distances <- absolute_copy_number_distance_grid(segments$copy_number,
                                                segments$weight,
                                                distance_function = "MAD")
distance_heatmap(distances)
}
