% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{find_minimum}
\alias{find_minimum}
\title{Find local minimum in distance function on fitting relative copy numbers}
\usage{
find_minimum(
  ploidy,
  cellularity,
  relative_copy_numbers,
  weights = NULL,
  distance_function = c("MAD", "RMSD")
)
}
\arguments{
\item{ploidy}{the tumour ploidy.}

\item{cellularity}{the cellularity, i.e. the fraction of cells that are from
the tumour.}

\item{relative_copy_numbers}{a numeric vector containing relative copy
numbers, i.e. ratios of copy numbers to the average copy number.}

\item{weights}{a numeric vector of weights to apply to each copy number value
(should be same length as relative_copy_numbers)}

\item{distance_function}{the distance function to use, either "MAD" for the
mean absolute difference or "RMSD" for the root mean square difference, where
differences are between the fitted absolute copy number values and the
nearest whole number.}
}
\value{
the distance in the fitted absolute copy numbers to whole numbers.
}
\description{
Finds the local minimum in the distance function from scaling relative copy
numbers to absolute copy number starting from the given ploidy and
cellularity.
}
\examples{
data(copy_number)
copy_number <- copy_number[copy_number$sample == "X17222", ]
find_minimum(3, 0.67, copy_number$segmented)
}
