% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{genes}
\alias{genes}
\title{Sample set of genes}
\format{
A data frame with 4 rows and 4 variables:
\describe{
  \item{name}{the gene name or symbol}
  \item{chromosome}{the chromosome}
  \item{start}{the start position of the gene within the chromosome}
  \item{end}{the end position of the gene within the chromosome}
}
}
\usage{
genes
}
\description{
A set of genes and their locations to illustrate display of genes in the
chromosome copy number plot.
}
\keyword{datasets}
