% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{is_acceptable_ploidy_and_cellularity}
\alias{is_acceptable_ploidy_and_cellularity}
\title{Determine if a given ploidy and cellularity for fitting copy numbers is
acceptable}
\usage{
is_acceptable_ploidy_and_cellularity(
  ploidy,
  cellularity,
  relative_copy_numbers,
  weights = NULL,
  max_proportion_zero = 0.05,
  min_proportion_close_to_whole_number = 0.5,
  max_distance_from_whole_number = 0.15
)
}
\arguments{
\item{ploidy}{the tumour ploidy.}

\item{cellularity}{the cellularity, i.e. the fraction of cells that are from
the tumour.}

\item{relative_copy_numbers}{a numeric vector containing relative copy
numbers, i.e. ratios of copy numbers to the average copy number.}

\item{weights}{a numeric vector of weights to apply to each copy number value
(should be same length as relative_copy_numbers)}

\item{max_proportion_zero}{the maximum proportion of fitted absolute copy
number values in the zero copy number state.}

\item{min_proportion_close_to_whole_number}{the minimum proportion of fitted
absolute copy number values sufficiently close to a whole number.}

\item{max_distance_from_whole_number}{the maximum distance from a whole
number that a fitted absolute copy number can be to be considered
sufficiently close.}
}
\description{
Determines if the given ploidy and cellularity are compatible with the
provided relative copy numbers based on the proportion of copy number values
that correspond to or are below the zero copy number state and on the
proportion that are sufficiently close to a whole number copy number state.
}
\examples{
data(copy_number)
copy_number <- copy_number[copy_number$sample == "X17222", ]
is_acceptable_ploidy_and_cellularity(3, 0.67, copy_number$segmented)
}
