% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{relative_to_absolute_copy_number}
\alias{relative_to_absolute_copy_number}
\title{Convert relative copy numbers to absolute copy numbers}
\usage{
relative_to_absolute_copy_number(relative_copy_numbers, ploidy, cellularity)
}
\arguments{
\item{relative_copy_numbers}{a numeric vector containing relative copy
numbers, i.e. ratios of copy numbers to the average copy number.}

\item{ploidy}{the tumour ploidy.}

\item{cellularity}{the cellularity, i.e. the fraction of cells that are
from the tumour.}
}
\value{
a numeric vector of absolute copy numbers.
}
\description{
Convert relative copy numbers to absolute copy numbers based on the given
ploidy and cellularity.
}
\examples{
relative_to_absolute_copy_number(c(0.98, 1.6, 1.23), 4.01, 0.77)
}
