% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{genome_copy_number_plot}
\alias{genome_copy_number_plot}
\title{Copy number plot for the whole genome}
\usage{
genome_copy_number_plot(
  copy_number,
  segments,
  sample = NULL,
  chromosome_lengths = NULL,
  copy_number_steps = NULL,
  max_points_to_display = Inf,
  min_copy_number = NULL,
  max_copy_number = NULL,
  copy_number_breaks = NULL,
  point_colour = "black",
  point_alpha = 0.15,
  point_size = 0,
  segment_colour = "red",
  segment_alpha = 1,
  segment_line_size = 0.75,
  copy_number_step_colour = "blue",
  copy_number_step_alpha = 0.35,
  copy_number_step_line_size = 0.75,
  xlabel = "chromosome",
  ylabel = "copy number"
)
}
\arguments{
\item{copy_number}{a data frame containing \code{chromosome}, \code{start},
\code{end} and \code{copy_number} columns where there is a row for each copy
number bin; may optionally contain a \code{sample} column if the data frame
contains data for multiple samples.}

\item{segments}{a data frame containing \code{chromosome}, \code{start},
\code{end} and \code{copy_number} columns; may optionally contain a
\code{sample} column if the data frame contains data for multiple samples.}

\item{sample}{the sample (required if the \code{copy_number} and
\code{segments} contain data for multiple samples)}

\item{chromosome_lengths}{a data frame containing \code{chromosome} and
\code{length} columns (optional).}

\item{copy_number_steps}{a data frame containing \code{absolute_copy_number}
and \code{copy_number} columns.}

\item{max_points_to_display}{maximum number of copy number points to display
(downsampling may be carried out if there are more copy number values than
this number).}

\item{min_copy_number}{the minimum \code{copy_number} to display.}

\item{max_copy_number}{the maximum \code{copy_number} to display.}

\item{copy_number_breaks}{breaks at which grid lines will be displayed.}

\item{point_colour}{the colour of the copy number points.}

\item{point_alpha}{the transparency of the copy number points.}

\item{point_size}{= the size of the copy number points.}

\item{segment_colour}{the colour of the copy number segments.}

\item{segment_alpha}{the transparency of the copy number segments.}

\item{segment_line_size}{the size of the lines for copy number segments.}

\item{copy_number_step_colour}{the colour of the copy number step lines.}

\item{copy_number_step_alpha}{the transparency of the copy number step lines.}

\item{copy_number_step_line_size}{the size of the lines for the copy number
steps.}

\item{xlabel, ylabel}{x- and y-axis labels.}
}
\value{
a \code{ggplot} object
}
\description{
Whole genome copy number plot with partitioning lines between chromosomes.
}
\details{
\code{copy_number} values can be on the relative or absolute scales or can
be log2 ratios but the same scale should be used consistently in each of the
\code{copy_number}, \code{segments} and \code{copy_number_steps} data frames.
}
\examples{
data(copy_number)

segments <- copy_number_segments(copy_number)

genome_copy_number_plot(copy_number, segments, sample = "X17222", ylabel = "relative copy number")

absolute_copy_numbers <- 0:8
relative_copy_numbers <- absolute_to_relative_copy_number(absolute_copy_numbers, ploidy = 4.01, cellularity = 0.81)
copy_number_steps <- data.frame(absolute_copy_number = absolute_copy_numbers, copy_number = relative_copy_numbers)

genome_copy_number_plot(
  copy_number,
  segments,
  sample = "X17222",
  copy_number_steps = copy_number_steps,
  min_copy_number = 0.25,
  max_copy_number = 2.5,
  xlabel = NULL,
  ylabel = "relative copy number")

# filter for specific sample and convert relative copy numbers to log2 ratios
log2_ratio <- copy_number[copy_number$sample == "X17222", ]
log2_ratio$copy_number <- log2(log2_ratio$copy_number)
log2_ratio$segmented <- log2(log2_ratio$segmented)

log2_ratio_segments <- copy_number_segments(log2_ratio)

log2_ratio_steps <- copy_number_steps
log2_ratio_steps$copy_number <- log2(log2_ratio_steps$copy_number)

genome_copy_number_plot(
  log2_ratio,
  log2_ratio_segments,
  copy_number_steps = log2_ratio_steps,
  min_copy_number = -2,
  max_copy_number = 3,
  xlabel = NULL,
  ylabel = expression(log[2]~ratio))

# filter for specific sample and convert relative copy numbers to absolute copy numbers
absolute_copy_number <- copy_number[copy_number$sample == "X17222", ]
absolute_copy_number$copy_number <- relative_to_absolute_copy_number(absolute_copy_number$copy_number, ploidy = 4.01, cellularity = 0.81)
absolute_copy_number$segmented <- relative_to_absolute_copy_number(absolute_copy_number$segmented, ploidy = 4.01, cellularity = 0.81)

absolute_segments <- copy_number_segments(absolute_copy_number)

genome_copy_number_plot(
  absolute_copy_number,
  absolute_segments,
  min_copy_number = 0,
  max_copy_number = 10,
  copy_number_breaks = 0:10,
  ylabel = "absolute copy number") +
  ggplot2::theme(panel.grid.major.y = ggplot2::element_line(colour = "grey60"))

}
