% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{start_shiny_app}
\alias{start_shiny_app}
\title{Shiny app for absolute copy number scaling}
\usage{
start_shiny_app(
  load_sample_data = TRUE,
  launch_browser = TRUE,
  port = getOption("shiny.port")
)
}
\arguments{
\item{load_sample_data}{load a sample data set on starting the Shiny app.}

\item{launch_browser}{launch the application in a web browser on starting the
Shiny app.}

\item{port}{the TCP port that the application should listen on.}
}
\value{
opens the Shiny app in a web browser.
}
\description{
Starts the shiny app in a new tab in the the default web browser.
If a browser tab isn't opened it may be necessary to manually open the
URL shown in the console.
}
